# Conflicts:
#	app/src/main/java/apc/appcradle/kotlinjc_friendsactivity_app/ui/screens/main/PermittedUi.kt

+++ app/src/main/java/apc/appcradle/kotlinjc_friendsactivity_app/domain/model/AppSavedSettingsData.kt
@@ -3,5 +3,5 @@ package apc.appcradle.kotlinjc_friendsactivity_app.domain.model
 data class AppSavedSettingsData(
     val savedTheme: AppThemes = AppThemes.System,
     val savedUserStep: Double = 0.4,
-    val savedScale: Int = 1
+    val savedScale: Float = 1.0f
 )
